+++ core/util.py
@@ -1,6 +1,9 @@
 import yaml
 from glob import glob
 import os
+import re
+
+re_sp = re.compile(r"\s+")
 
 def chunks(lst, n):
     arr = []
@@ -31,3 +34,26 @@ def readlines(*fls):
                     i = i.strip()
                     if i:
                         yield i
+
+def parse_tag(_tag):
+    _tag = re_sp.sub(" ", _tag).strip()
+    if not _tag:
+        return None
+    tags = _tag.split()
+    if len(tags)>1:
+        tags=[parse_tag(t) or t for t in tags]
+        return " ".join(tags)
+    tag = _tag
+    for a, b in (
+        ("á", "a"),
+        ("é", "e"),
+        ("í", "i"),
+        ("ó", "o"),
+        ("ú", "u")
+    ):
+        tag = tag.replace(a, b)
+    if tag == "españa":
+        return "España"
+    if tag == "Europa":
+        return "Europa"
+    return _tag
